There are more than just trees, Gloria.
