+++ phpunit.xml.dist
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit backupGlobals="false" backupStaticAttributes="false" bootstrap="vendor/autoload.php" colors="true" convertErrorsToExceptions="true" convertNoticesToExceptions="true" convertWarningsToExceptions="true" processIsolation="false" stopOnFailure="false">
     <php>
-        <env name="TREE_IP" value="192.168.1.1"/>
+        <env name="DEVICE_IP" value="192.168.1.1"/>
     </php>
 </phpunit>
